Update README.md
Theoretical Approach and Method
All experiments will be carried out on DNA from cell lines to be purchased from the Coriell Institute. Different patient cells with lar Autosomal Dominant Spinoserebellar Ataxia (ODSCA), Fragile-X, Huntington Disease will be used because of high genetic heterogeneity. The loci known to be affected in the respective disease will be separated intact as much as possible from the remainder of the genome. For this, recently developed CRISPR / Cas9 mediated chromosomal enrichment approaches will be adopted (Gabrieli et al. 2017, 2018; Bennett-Baker et al. 2017). The important advantages of this approach are that CRISPR / Cas9 endonuclease system can be directed to the desired genomic target (s) easily and cheaply, to target more than one genomic locus in the same reaction, and to have such a high specificity. In the few studies reported in the literature to date, methods that require long-lasting and special equipment (eg, Pulse-field gel electrophoresis), and the target DNA was amplified with PCR or MDA (Gabrieli et al. 2018), or non-optimal steps in defining repetitive and complex structural variants, such as DNA sequencing by enrichment with CRISPR / Cas9 (Bennett-Baker et al. 2017). With the CRISPR / Cas9-mediated method to be developed in the proposed project, a much faster and more practical method will be developed, and it will be respond to clinical needs by targeting more than one locus in the same reaction.